refactor: extract template/dashboard-view/bundle routes into routes/templatesViews.ts
Pure structural move out of createApp; no handler logic changed. No RouteContext
members graduated (every handler reads its store off options.*).
